improve(altstorage):  use a consistent method for writing reports to storage
