+++ deploy/helm/templates/deployment.yaml
@@ -62,6 +62,8 @@ spec:
               value: "true"
             - name: OPERATOR_ALTERNATE_REPORT_STORAGE_DIR
               value: {{ .Values.alternateReportStorage.mountPath | quote }}
+            - name: OPERATOR_ALTERNATE_REPORT_STORAGE_PRETTY_PRINT
+              value: {{ .Values.alternateReportStorage.prettyPrint | quote }}
           {{- end }}
           envFrom:
             - configMapRef:
