+++ docs/getting-started/installation/configuration.md
@@ -9,7 +9,7 @@ You can configure Trivy-Operator to control its behavior and adapt it to your ne
 | `OPERATOR_NAMESPACE`                                         | N/A                    | See [Install modes](#install-modes)                                                                                                                                                                                                                                                                                                                                  |
 | `OPERATOR_TARGET_NAMESPACES`                                 | N/A                    | See [Install modes](#install-modes)                                                                                                                                                                                                                                                                                                                                  |
 | `OPERATOR_EXCLUDE_NAMESPACES`                                | N/A                    | A comma separated list of namespaces (or glob patterns) to be excluded from scanning in all namespaces [Install mode](#install-modes).                                                                                                                                                                                                                               |
- | `OPERATOR_TARGET_WORKLOADS`                                  | All workload resources | A comma separated list of Kubernetes workloads to be included in the vulnerability and config-audit scans                                                                                                                                                                                                                                                            |
+| `OPERATOR_TARGET_WORKLOADS`                                  | All workload resources | A comma separated list of Kubernetes workloads to be included in the vulnerability and config-audit scans                                                                                                                                                                                                                                                            |
 | `OPERATOR_SERVICE_ACCOUNT`                                   | `trivy-operator`       | The name of the service account assigned to the operator's pod                                                                                                                                                                                                                                                                                                       |
 | `OPERATOR_LOG_DEV_MODE`                                      | `false`                | The flag to use (or not use) development mode (more human-readable output, extra stack traces and logging information, etc).                                                                                                                                                                                                                                         |
 | `OPERATOR_SCAN_JOB_TTL`                                      | `""`                   | The set automatic cleanup time after the job is completed                                                                                                                                                                                                                                                                                                            |
@@ -45,6 +45,7 @@ You can configure Trivy-Operator to control its behavior and adapt it to your ne
 | `OPERATOR_SBOM_GENERATION_ENABLED`                           | `true`                 | The flag to enable sbom generation                                                                                                                                                                                                                                                                                                                                   |
 | `OPERATOR_ALTERNATE_REPORT_STORAGE_ENABLED`                  | `false`                | Control where reports are written. By default this is false, so reports will be written normally as CRDs in ETCD memory. However, if you would rather reports be written to a persistent volume, flip this to true. If done a persistent volume claim will be inluded in your installation and all reports will be written there. See settings for more granularity. |
 | `OPERATOR_ALTERNATE_REPORT_STORAGE_DIR`                      | `/var/reports`         | Specifies the directory path for storing reports when `OPERATOR_ALTERNATE_REPORT_STORAGE_ENABLED` is set to `true`. Ensure that the specified directory is writable and has sufficient storage capacity.                                                                                                                                                             |
+| `OPERATOR_ALTERNATE_REPORT_STORAGE_PRETTY_PRINT`             | `false`                | The flag is to enable "pretty printed" JSON formatting when writing reports to the alternate storage.                                                                                                                                                                                                                                                                |
 The values of the `OPERATOR_NAMESPACE` and `OPERATOR_TARGET_NAMESPACES` determine the installation mode, which in turn determines the multitenancy support of the operator.
 
 | MODE            | OPERATOR_NAMESPACE | OPERATOR_TARGET_NAMESPACES | DESCRIPTION                                                                                                    |
