+++ pkg/rbacassessment/builder.go
@@ -2,7 +2,6 @@ package rbacassessment
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -22,6 +21,7 @@ import (
 	"github.com/aquasecurity/trivy-operator/pkg/kube"
 	"github.com/aquasecurity/trivy-operator/pkg/operator/etc"
 	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
+	"github.com/aquasecurity/trivy-operator/pkg/utils"
 )
 
 type ReportBuilder struct {
@@ -185,14 +185,10 @@ func (b *ReportBuilder) Write(ctx context.Context, writer Writer) error {
 			log.V(1).Info("Writing cluster RBAC assessment report to alternate storage", "dir", b.Config.AltReportDir)
 			// Write the cluster RBAC assessment report to a file
 			reportDir := b.Config.AltReportDir
+			reportPrettyPrint := b.Config.AltReportPrettyPrint
 			rbacAssessmentDir := filepath.Join(reportDir, "cluster_rbac_assessment_reports")
 			if err := os.MkdirAll(rbacAssessmentDir, 0o750); err != nil {
 				return fmt.Errorf("failed to make rbacAssessmentDir %s: %w", rbacAssessmentDir, err)
-			}
-
-			reportData, err := json.Marshal(report)
-			if err != nil {
-				return fmt.Errorf("failed to marshal cluster RBAC assessment report: %w", err)
 			}
 
 			// Extract workload kind and name from report labels
@@ -200,7 +196,7 @@ func (b *ReportBuilder) Write(ctx context.Context, writer Writer) error {
 			workloadName := report.Labels["trivy-operator.resource.name"]
 
 			reportPath := filepath.Join(rbacAssessmentDir, fmt.Sprintf("%s-%s.json", workloadKind, workloadName))
-			err = os.WriteFile(reportPath, reportData, 0o600)
+			err = utils.StreamReportToFile(report, reportPath, 0o600, reportPrettyPrint)
 			if err != nil {
 				return fmt.Errorf("failed to write cluster RBAC assessment report: %w", err)
 			}
