+++ pkg/vulnerabilityreport/controller/scanjob.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -25,28 +24,11 @@ import (
 	"github.com/aquasecurity/trivy-operator/pkg/operator/etc"
 	"github.com/aquasecurity/trivy-operator/pkg/sbomreport"
 	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
+	"github.com/aquasecurity/trivy-operator/pkg/utils"
 	"github.com/aquasecurity/trivy-operator/pkg/vulnerabilityreport"
 
 	. "github.com/aquasecurity/trivy-operator/pkg/operator/predicate"
 )
-
-// streamReportToFile writes a report directly to file using streaming JSON encoding
-// This reduces memory usage by avoiding intermediate marshaling to byte arrays
-func streamReportToFile(report any, filePath string) error {
-	file, err := os.Create(filePath)
-	if err != nil {
-		return fmt.Errorf("failed to create file %s: %w", filePath, err)
-	}
-	defer file.Close()
-
-	encoder := json.NewEncoder(file)
-	encoder.SetIndent("", "  ") // Pretty print for readability
-	if err := encoder.Encode(report); err != nil {
-		return fmt.Errorf("failed to encode report: %w", err)
-	}
-
-	return nil
-}
 
 // ScanJobController watches Kubernetes workloads and generates
 // v1alpha1.VulnerabilityReport instances using vulnerability scanner that that
@@ -352,6 +334,7 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context,
 		log.V(1).Info("Writing vulnerability reports to alternate storage", "dir", r.Config.AltReportDir)
 		// Initilizing alternate writing directory var for trivy reports
 		reportDir := r.Config.AltReportDir
+		reportPrettyPrint := r.Config.AltReportPrettyPrint
 		// Create subdirectories for each type of report
 		clusterVulnerabilityDir := filepath.Join(reportDir, "cluster_vulnerability_reports")
 		vulnerabilityDir := filepath.Join(reportDir, "vulnerability_reports")
@@ -372,7 +355,7 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context,
 		// Write cluster vulnerability reports to a file using streaming
 		if vulnerabilityReports.vulnerabilityClusterReports != nil {
 			reportPath := filepath.Join(clusterVulnerabilityDir, fmt.Sprintf("%s-%s-%s.json", workloadKind, workloadName, containerName))
-			if err := streamReportToFile(vulnerabilityReports.vulnerabilityClusterReports, reportPath); err != nil {
+			if err := utils.StreamReportToFile(vulnerabilityReports.vulnerabilityClusterReports, reportPath, 0o644, reportPrettyPrint); err != nil {
 				return VulnerabilityReports{}, []v1alpha1.ExposedSecretReport{}, &SbomReports{}, fmt.Errorf("failed to write cluster vulnerability report: %w", err)
 			}
 		}
@@ -380,7 +363,7 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context,
 		// Write vulnerability reports to a file using streaming
 		if vulnerabilityReports.vulnerabilityNamespaceReports != nil {
 			reportPath := filepath.Join(vulnerabilityDir, fmt.Sprintf("%s-%s-%s.json", workloadKind, workloadName, containerName))
-			if err := streamReportToFile(vulnerabilityReports.vulnerabilityNamespaceReports, reportPath); err != nil {
+			if err := utils.StreamReportToFile(vulnerabilityReports.vulnerabilityNamespaceReports, reportPath, 0o644, reportPrettyPrint); err != nil {
 				return VulnerabilityReports{}, []v1alpha1.ExposedSecretReport{}, &SbomReports{}, fmt.Errorf("failed to write vulnerability report: %w", err)
 			}
 		}
@@ -388,7 +371,7 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context,
 		// Write secret reports to a file using streaming
 		if len(secretReports) > 0 {
 			reportPath := filepath.Join(secretDir, fmt.Sprintf("%s-%s-%s.json", workloadKind, workloadName, containerName))
-			if err := streamReportToFile(secretReports, reportPath); err != nil {
+			if err := utils.StreamReportToFile(secretReports, reportPath, 0o644, reportPrettyPrint); err != nil {
 				return VulnerabilityReports{}, []v1alpha1.ExposedSecretReport{}, &SbomReports{}, fmt.Errorf("failed to write exposed secrets report: %w", err)
 			}
 		}
@@ -396,7 +379,7 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context,
 		// Write SBOM cluster reports to a file using streaming
 		if len(sbomReports.sbomClusterReports) > 0 {
 			reportPath := filepath.Join(clusterSbomDir, fmt.Sprintf("%s-%s-%s.json", workloadKind, workloadName, containerName))
-			if err := streamReportToFile(sbomReports.sbomClusterReports, reportPath); err != nil {
+			if err := utils.StreamReportToFile(sbomReports.sbomClusterReports, reportPath, 0o644, reportPrettyPrint); err != nil {
 				return VulnerabilityReports{}, []v1alpha1.ExposedSecretReport{}, &SbomReports{}, fmt.Errorf("failed to write sbom cluster report: %w", err)
 			}
 		}
@@ -404,7 +387,7 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context,
 		// Write SBOM reports to a file using streaming
 		if len(sbomReports.sbomNamespaceReports) > 0 {
 			reportPath := filepath.Join(sbomDir, fmt.Sprintf("%s-%s-%s.json", workloadKind, workloadName, containerName))
-			if err := streamReportToFile(sbomReports.sbomNamespaceReports, reportPath); err != nil {
+			if err := utils.StreamReportToFile(sbomReports.sbomNamespaceReports, reportPath, 0o644, reportPrettyPrint); err != nil {
 				return VulnerabilityReports{}, []v1alpha1.ExposedSecretReport{}, &SbomReports{}, fmt.Errorf("failed to write sbom report: %w", err)
 			}
 		}
